split out timerfd logic into separate 'TimerFDCtx'
