+++ src/eventfd_ctx.h
@@ -17,9 +17,7 @@ errno_t eventfd_ctx_init(EventFDCtx *eventfd, int kq, unsigned int counter,
     int flags);
 errno_t eventfd_ctx_terminate(EventFDCtx *eventfd);
 
-int eventfd_ctx_fd(EventFDCtx *eventfd);
-
 errno_t eventfd_ctx_write(EventFDCtx *eventfd, int kq, uint64_t value);
-errno_t eventfd_ctx_read(EventFDCtx *eventfd, int kq_, uint64_t *value);
+errno_t eventfd_ctx_read(EventFDCtx *eventfd, int kq, uint64_t *value);
 
 #endif
